Use macros for safe multiplication.
Having functions in a separate file is a real performance hit. I would
like to use the builtin functions of gcc, but that would make it harder
to port to systems using older versions or other compilers. And I
would have to figure out the correct size of size_t on such systems...

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>