Greatly improve performance on large images.
If not all outputs are set, only the modified regions have to be sent.
This could also be done with --maximize and --center by figuring out
where they end up by hand instead of using pixman's translate for that.

If the image is too large for one request, don't waste time by
calling subimage repeatedly. The scanlines have the advantage that the
data pointer can be simply increased. In case this is not true, please
contact me. Haven't found a reason why this shouldn't work.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>