Introduced DBG for debug messages.
Getting rid of all these #ifdef's is really needed to keep this readable
in any form.

The debug message are always interpreted by the compiler now, which is a
good thing. If variables change, the compiler will inform about it, not
just in the next rare moments somebody compiles with --with-debug.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>