Released version 0.4.1.
The --daemon option has been fixed to properly work with sandboxing.
The process is split early on so no further system calls have been
added to stage 2.

This version adds runtime debug support to display debug messages on
standard error output.  This is the same as the previous debug
configure option during build.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>