Adjusted dependency to libjpeg-turbo.
With this change, CMYK is gone as well. If you really have a CMYK JPEG
file which you want to use as a wallpaper, you are most likely trained
enough to be annoyed by a poor CMYK<->RGB conversion without any form
of color profile adjustments. If not, then you don't need a CMYK JPEG
either. Convert to RGB or Grayscale and everything's working again.

Besides, the CMYK<->RGB conversion in the code was rather poor. So this
is clearly no loss.

The biggest advantage is a speed-up in RGB and Grayscale parsing. When
using libjpeg-turbo it is possible to convert directly into the needed
RGBA without having a manual for-loop. Gained a 5% speed increasement
on my system.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>