Added --trim option.
A trim box, as known from preflight processing and PDF specs,
specifies the area of input files which are supposed to be
processed.

It is not a crop box, i.e. the input files are not cut down.
Instead, if output cannot be completely filled with content of
the trim box, additionally required pixels are retrieved from
input file.

This allows a nicer experience of --maximize and screens in
different dimensions.

Also, a trim box of size 1x1 allows to put a new center of image
for all different output modes.

A great feature, based on a request by Marc Espie.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>