Implemented new mode: focus
It is a mode that works best in conjuction with --trim.

With trim, specify the area of interest in your image file.
This area will always be visible on the output with --focus.

The trim area will only be zoomed in if this is the only way
to avoid black pixels. Otherwise the image will be moved
on output to make sure that as many output pixels are covered
with input pixels.

This feature is great if you have a single input file which
1) you want to use on multiple outputs with different aspect ratios
2) zooming into the center of the image is not an attractive choice.

Imagine having a portrait wallpaper sufficiently large for your
landscape output. You could center or zoom in, but without --focus
you cannot state that you want to center as good as possible and
only zoom in if needed.

It would be quite tedious to repeat this for every output device
or when output changes, i.e. a virtual machine with window resizing.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>