Always redraw on XRandR events.
User "themad" on GitHub pointed out that modifying the order of
displays does not lead to proper redrawing of wallpapers in daemon
mode.

The code only composed wallpapers again if dimensions changed. The idea
was to keep calculation overhead low, since XRandR events are sent
multiple times -- which makes sense.

If positions of displays are swapped, every display would inform about
this event. I could store somewhere if I have recomposed the wallpaper
due to one switch, but I would have to store offsets, sizes, etc. just
for a slight performance gain in an event in which displays are
modified.

These few milliseconds are not worth it.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>