Improved atom handling.
Atoms are used in xwallpaper to support pseudo transparency. Until now
it has been set to the same pixmap as drawn to root window.

Not every setup requires support for pseudo transparency. In fact it is
probably only used by terminal emulators which retrieve background
through this feature.

Now it is even possible to update the background used for pseudo
transparency independently from actual wallpaper. You can use this to
improve pseudo transparency for clients which do not support image
manipulation. Example would be to create a second image which is
darker than the first one. A terminal emulator which cannot reduce
brightness on its own would profit from this.

By default xwallpaper will try to retrieve the background from pseudo
transparency atoms. To stay with old behavior, always use --clear.

Switching xwallpaper behavior for scenarios in which not all outputs are
specified is a "corner case". It led to confusion in the past. Although
the whole X background setup is confusing, xwallpaper should try to
behave as (intuitively) expected.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>