Kill resources, not clients
The atom handling is a bit more complicated than expected. If --clear is
not specified, we do want to re-use atoms. This in turn means that we
will not create a new pixmap, which in turn means that we have no
resources which should retain.

We do want to retain pixmaps permanently if they are created and used in
atoms.

This in turn means that daemon mode is a bit more complicated. We do
want to re-use pixmaps of other programs (or xwallpaper instances) if we
can re-use them. This is not true anymore if the resolution of the
screen changed.

The simplicity of xwallpaper suffered greatly with separated root window
and atom handling as well as the daemon mode. This code is prone for
some refactoring.

For now it has to withstand further testing to see if my understanding
of pixmap handling in root windows and atoms is correct and if this in
turn resulted in a proper implementation.

Reported by and resolved with stharward on github.

Please note:

We still have an issue if daemon mode gets interrupted by another
wallpaper setter or by xwallpaper --clean. The allocated pixmap is
permanently retained and will not be cleaned up if the daemon xwallpaper
has not been explicitly killed before.

Since this sounds like an uncommon setup, I delay the fix. It might
involve adjusting the stage 2 sandbox, so I am willing to accept this
corner case -- documented in manual page.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>